requirements.txt
gymnasium>=0.29
highway-env>=1.8.2
numpy>=1.24
Pillow>=9.5
matplotlib>=3.7
joblib>=1.3
tqdm>=4.66
torch>=2.1
torchvision>=0.16
transformers>=4.34
stable-baselines3>=2.2
